requirements.txt
oauthenticator
sqlalchemy
configurable-http-proxy
setuptools
wheel
jupyterhub-idle-culler
jupyter_server
notebook
tornado
pycurl
pyjwt
mwoauth
jupyterhub-dummyauthenticator
globus_sdk[jwt]
jupyterhub-kubespawner
kubernetes
setuptools
Twisted
Django
lxml
PyJWT
cryptography
ruamel.yaml